// api_services/auth/auth.interface.ts
export type SendOtpDto = { auth_param: string | number | null };
export type SendForgetOtpDto = { auth_param: string | number | null; forget_type: string | null };
export type ConfirmForgetOtpDto = {
  auth_param: string | number | null;
  forget_type: string | null;
  code: string | null;
};
export type SendOtpType = { code: number };
export type SendOtpVerify = { auth_param: number | null | string; code: number | null | string };
export type SendOtpVerifyResponse = any;
export type CheckUserName = { username: string };
export type CheckUserNameResponse = any;
export type SetPassword = { username?: string; password: string | null; password_confirm: string | null };
export type SetNewPassword = {
  password: string | null;
  password_confirm: string | null;
  auth_param: string | null;
  forget_type: string;
  code: string;
};
export type SetDetailResponse = any;
export type SetDetail = {
  birthday: { year: string | number | null; month: string | number | null; day: string | number | null } | null;
  profile_image_id: string | number | null;
  city_id: string | number | null;
  full_name: string | number | null;
};
export type SetPasswordResponse = any;
export type SignInDTO = {
  password: string | number | null;
  auth_param: string | number | null;
};
export type SignInResponseDTO = any;

// Generated by https://quicktype.io

export interface ForgetasswordListDto {
  [key: string]: string | null;
}

export interface GetProfileDto {
  id: number;
  first_name: string | null;
  last_name: string | null;
  mobile_number: string;
  email: string | null;
  birth_date: string | null;
  gender: "Mrs" | "Mr" | "Ms" | null;
  credit: number;
  zip_codes: Array<string[]>;
  created_at: string;
  updated_at: string;
  email_verified_at: null;
  mobile_verified_at: string;
  image_location?: string;
}

export type UpdateProfileDto = {
  first_name: string;
  last_name: string;
  birth_date: string;
  gender: "Mrs" | "Mr" | "Ms" | null;
};

// Generated by https://quicktype.io